feat: Подключил стартовое модальное окно на странице игры
