+++ packages/client/src/pages/game/components/GamePage/index.ts
@@ -1,1 +1,1 @@
-export { GamePage } from './GamePage'
+export { GamePage, initGamePage } from './GamePage'
