+++ packages/client/src/routes/routes.tsx
@@ -5,7 +5,7 @@ import { initLoginPage, LoginPage } from '../pages/login/Login'
 import { ProtectedRoute } from './ProtectedRoute'
 import { initProfilePage, ProfilePage } from '../pages/profile/ProfilePage'
 import { AppErrorBoundary } from '../components/AppErrorBoundary'
-import { GamePage, initGamePage } from '../pages/Game'
+import { GamePage, initGamePage } from '../pages/game/components/GamePage'
 import { Error500Page, initError500Page } from '../pages/error500/Error500'
 import { Error404Page, initError404Page } from '../pages/error404/Error404'
